upd: replace feat

+++ index.js
@@ -242,42 +242,42 @@ app.post("/api/update", (req, res) => {
 
 app.post("/api/updateDB", async (req, res) => {
   var newData = req.body;
-  await connection.query(
-    "DELETE FROM practitioner_list",
-    (error, results, fields) => {}
-  );
+  if (newData.replace) {
+    await connection.query(
+      "DELETE FROM practitioner_list",
+      (error, results, fields) => {}
+    );
+    // Add Admin
+    await connection.query(
+      "INSERT INTO practitioner_list (firstname, lastname, specialty, imageURL, upload, tags, meetinglink, address, city, state, zipcode, country, email, phone, `rank`, review, status, role, password, sex) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)",
+      [
+        "Nima",
+        "Farshid",
+        "Bio-Well practitioner, Reiki Master, Sound Healer, Meditation Coach",
+        "https://biohackingcongress.com/storage/users/June2023/9Q67Ebbs5rPLWWmWGZET.png",
+        0,
+        "Reiki, biowell, soundhealer, meditation",
+        "https://calendly.com/nimafarshid/biowell",
+        "11532 Via Lucerna Cir",
+        "Windermere",
+        "FL",
+        "34786",
+        "US",
+        "nima02@yahoo.com",
+        "407-230-8179",
+        3,
+        5,
+        "active",
+        0,
+        "$2b$10$WZ9pp7nsSEcgglZD8W8oueFvDfSDKKY1VJ.wVWRGRKubqDlowH2UG",
+        "Male",
+      ],
+      (error, results, fields) => {
+        if (error) throw error;
+      }
+    );
+  }
 
-  // Add Admin
-  await connection.query(
-    "INSERT INTO practitioner_list (firstname, lastname, specialty, imageURL, upload, tags, meetinglink, address, city, state, zipcode, country, email, phone, `rank`, review, status, role, password, sex) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)",
-    [
-      "Nima",
-      "Farshid",
-      "Bio-Well practitioner, Reiki Master, Sound Healer, Meditation Coach",
-      "https://biohackingcongress.com/storage/users/June2023/9Q67Ebbs5rPLWWmWGZET.png",
-      0,
-      "Reiki, biowell, soundhealer, meditation",
-      "https://calendly.com/nimafarshid/biowell",
-      "11532 Via Lucerna Cir",
-      "Windermere",
-      "FL",
-      "34786",
-      "US",
-      "nima02@yahoo.com",
-      "407-230-8179",
-      3,
-      5,
-      "active",
-      0,
-      "$2b$10$WZ9pp7nsSEcgglZD8W8oueFvDfSDKKY1VJ.wVWRGRKubqDlowH2UG",
-      "Male",
-    ],
-    (error, results, fields) => {
-      if (error) throw error;
-    }
-  );
-
-  console.log(newData);
   // Add users
   newData.forEach(async (element, index) => {
     if (element["Email"] != "nima02@yahoo.com" && index != 0) {
